pyproject.toml
[tool.poetry]
name = "lensor-case-study"
version = "0.1.0"
description = ""
authors = ["heliyahasani <heliya.hasani@fbistech.com>"]
readme = "README.md"

packages = [{include="*", from="src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
numpy = ">=1.23.2"
pytest = "^8.3.2"
prefect = "^2.20.0"
tqdm = "^4.66.5"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.5"
tb-nightly = "^2.18.0a20240807"

[tool.poetry.group.lint.dependencies]
pylint = "^3.0.2"
isort = "^5.12.0"
black = "^23.10.1"
autoflake = "^2.2.1"

[tool.black]
line-length=120

[tool.isort]
profile = "black"

[tool.autoflake]
remove-all-unused-imports = true
ignore-init-module-imports = true
quiet = true
in-place = true
recursive = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
